tgbotpy/requirements.txt
python-telegram-bot==20.4
opencv-python>=4.7.0.72
